requirements.txt
ipython>=8.10
jupyterlab>=3.0
kedro-datasets[pandas-csvdataset, pandas-exceldataset, pandas-parquetdataset, spark-sparkdataset, plotly-plotlydataset, plotly-jsondataset, matplotlib-matplotlibwriter]>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.12
notebook
scikit-learn~=1.5.1
seaborn~=0.12.1
setuptools; python_version >= "3.10"
pandas
autogluon.tabular>=0.8.0
pyspark>=3.5.5
torch>=2.7.0